+++ src/main/java/com/example/springblog/models/User.java
@@ -3,36 +3,31 @@ package com.example.springblog.models;
 import javax.persistence.Column;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
-import javax.persistence.GenerationType;
 import javax.persistence.Id;
-import javax.persistence.OneToOne;
-import javax.validation.constraints.Pattern;
-import javax.validation.constraints.Size;
-import org.hibernate.validator.constraints.NotBlank;
 
 @Entity
 public class User {
     @Id
-    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    @GeneratedValue
     private long id;
 
-    @NotBlank(message = "First name can not be blank")
-    @Size(min = 3, message = "First name is required")
+//    @NotBlank(message = "First name can not be blank")
+//    @Size(min = 3, message = "First name is required")
     @Column(nullable = false, length = 100)
     private String first_name;
 
-    @NotBlank(message = "Last name can not be blank")
-    @Size(min = 3, message = "Last name is required")
+//    @NotBlank(message = "Last name can not be blank")
+//    @Size(min = 3, message = "Last name is required")
     @Column(nullable = false, length = 100)
     private String last_name;
 
-    @Column(nullable = false, length = 1000)
+    @Column(nullable = true, length = 1000)
     private String address1;
 
-    @Column(nullable = false, length = 1000)
+    @Column(nullable = true, length = 1000)
     private String address2;
 
-    @Column(nullable = false, length = 500)
+    @Column(nullable = true, length = 500)
     private String city;
 
     @Column(nullable = false, length = 2)
@@ -41,33 +36,30 @@ public class User {
     @Column(nullable = false)
     private String postalCode;
 
-    @NotBlank(message = "Email can not be blank")
-    @Pattern(regexp = "^[A-Za-z0-9+_.-]+@(.+)$", message = "Email must be formatted correctly")
+//    @NotBlank(message = "Email can not be blank")
+//    @Pattern(regexp = "^[A-Za-z0-9+_.-]+@(.+)$", message = "Email must be formatted correctly")
     @Column(nullable = false, length = 100, unique = true)
     private String email;
 
     @Column(nullable = false, length = 100, unique = true)
     private String username;
 
-    @NotBlank(message = "Password can not be blank")
-    @Pattern(regexp = "^(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z])(?=.*[!@#$%^&+=])(?=\\S+$).{8,}$",
-        message = "Password must have one uppercase letter, one number, and one special "
-            + "character")
-    @Size(min = 8, message = "Password must be at least 8 characters")
+//    @NotBlank(message = "Password can not be blank")
+//    @Pattern(regexp = "^(?=.*[0-9])(?=.*[a-z])(?=.*[A-Z])(?=.*[!@#$%^&+=])(?=\\S+$).{8,}$",
+//        message = "Password must have one uppercase letter, one number, and one special "
+//            + "character")
+//    @Size(min = 8, message = "Password must be at least 8 characters")
     @Column(nullable = false, length = 100)
     private String password;
 
-    @NotBlank(message = "Phone number can not be blank")
-    @Size(min = 10, message = "Phone number required")
+//    @NotBlank(message = "Phone number can not be blank")
+//    @Size(min = 10, message = "Phone number required")
     @Column(nullable = false)
     private String phone_number;
 
-    @NotBlank(message = "Date of birth can not be blank")
+//    @NotBlank(message = "Date of birth can not be blank")
     @Column(nullable = false)
-    private long dob;
-
-    @OneToOne
-    private Role role;
+    private int dob;
 
     public User() {
     }
@@ -80,7 +72,7 @@ public class User {
 
     public User(String first_name, String last_name, String address1, String address2,
         String city, String state, String postalCode, String email, String username,
-        String password, String phone_number, long dob, Role role) {
+        String password, String phone_number, int dob) {
         this.first_name = first_name;
         this.last_name = last_name;
         this.address1 = address1;
@@ -93,23 +85,13 @@ public class User {
         this.password = password;
         this.phone_number = phone_number;
         this.dob = dob;
-        this.role = role;
     }
 
     public User(User copy) {
         id = copy.id;
-        first_name = copy.first_name;
-        last_name = copy.last_name;
-        address1 = copy.address1;
-        address2 = copy.address2;
-        city = copy.city;
-        state = copy.state;
-        postalCode = copy.postalCode;
         email = copy.email;
         username = copy.username;
         password = copy.password;
-        phone_number = copy.phone_number;
-        dob = copy.dob;
     }
 
     public long getId() {
@@ -212,15 +194,8 @@ public class User {
         return dob;
     }
 
-    public void setDob(long dob) {
+    public void setDob(int dob) {
         this.dob = dob;
     }
 
-    public Role getRole() {
-        return role;
-    }
-
-    public void setRole(Role role) {
-        this.role = role;
-    }
 }
