+++ src/main/java/com/example/springblog/models/UserRole.java
@@ -5,6 +5,7 @@ import javax.persistence.*;
 @Entity
 @Table(name = "user_roles")
 public class UserRole {
+
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private long id;
@@ -14,4 +15,33 @@ public class UserRole {
 
     @Column(name = "role")
     private String role;
+
+    public UserRole(long userId, String role) {
+        this.userId = userId;
+        this.role = role;
+    }
+
+    public long getId() {
+        return id;
+    }
+
+    public void setId(long id) {
+        this.id = id;
+    }
+
+    public long getUserId() {
+        return userId;
+    }
+
+    public void setUserId(long userId) {
+        this.userId = userId;
+    }
+
+    public String getRole() {
+        return role;
+    }
+
+    public void setRole(String role) {
+        this.role = role;
+    }
 }
