+++ src/main/java/com/example/springblog/models/UserWithRoles.java
@@ -12,17 +12,12 @@ public class UserWithRoles extends User implements UserDetails {
 
     public UserWithRoles(User user) {
         super(user);  // Call the copy constructor defined in User
-        this.userRole = user.getRole().getName();
     }
 
     @Override
     public Collection<? extends GrantedAuthority> getAuthorities() {
         String roles = ""; // Since we're not using the authorization part of the component
         return AuthorityUtils.commaSeparatedStringToAuthorityList(roles);
-    }
-
-    public String getUserRole() {
-        return userRole;
     }
 
     @Override
