+++ src/main/java/com/example/springblog/services/UserRepository.java
@@ -1,0 +1,9 @@
+package com.example.springblog.services;
+
+import com.example.springblog.models.User;
+import org.springframework.data.repository.CrudRepository;
+
+public interface UserRepository extends CrudRepository<User, Long> {
+  User findByUsername(String username);
+  User findById(Long id);
+}
